+++ modules/network/src/main/java/org/apache/ignite/internal/network/netty/ConnectionManager.java
@@ -23,7 +23,6 @@ import io.netty.bootstrap.Bootstrap;
 import java.net.InetSocketAddress;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -398,6 +397,6 @@ public class ConnectionManager implements ChannelCreationListener {
      */
     @TestOnly
     public Map<ConnectorKey<String>, NettySender> channels() {
-        return new HashMap<>(channels);
+        return Map.copyOf(channels);
     }
 }
